+++ src/main/kotlin/de/passbutler/desktop/base/ConfigProperties.kt
@@ -1,7 +1,7 @@
 package de.passbutler.desktop.base
 
 object ConfigProperty {
-    const val RECENT_VAULT = "recentVault"
+    const val RECENT_VAULTS = "recentVaults"
     const val THEME_TYPE = "themeType"
     const val LANGUAGE_CODE = "language"
     const val PREMIUM_KEY = "premiumKey"