+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Password unmasking icon for password fields
 - Language hint text when importing CSV from KeePassX 2 / KeePassXC to be sure deleted entries are recognized as such
 - Pressing escape key on overview clears filter now
+- Recent vault files menu
 
 ## Changed
 - Allow to copy item details of read-only items (do not disable them anymore) - the save and delete button is hidden instead of disabled to avoid user confusion